+++ src/main/java/mafiadelprimobanco/focusproject/MainApplication.java
@@ -2,6 +2,7 @@ package mafiadelprimobanco.focusproject;
 
 import javafx.application.Application;
 import javafx.stage.Stage;
+import mafiadelprimobanco.focusproject.controller.handlers.SceneHandler;
 
 import java.io.IOException;
